+++ ui/src/app/core/accessories/accessory-tile/accessory-tile.component.ts
@@ -9,7 +9,6 @@ import { CarbonDioxideSensorComponent } from '@/app/core/accessories/types/carbo
 import { CarbonMonoxideSensorComponent } from '@/app/core/accessories/types/carbon-monoxide-sensor/carbon-monoxide-sensor.component'
 import { ContactSensorComponent } from '@/app/core/accessories/types/contact-sensor/contact-sensor.component'
 import { DoorComponent } from '@/app/core/accessories/types/door/door.component'
-import { FanV2Component } from '@/app/core/accessories/types/fan-v2/fan-v2.component'
 import { FanComponent } from '@/app/core/accessories/types/fan/fan.component'
 import { FilterMaintenanceComponent } from '@/app/core/accessories/types/filter-maintenance/filter-maintenance.component'
 import { GarageDoorOpenerComponent } from '@/app/core/accessories/types/garage-door-opener/garage-door-opener.component'
@@ -46,7 +45,6 @@ import { WindowComponent } from '@/app/core/accessories/types/window/window.comp
     ThermostatComponent,
     OutletComponent,
     FanComponent,
-    FanV2Component,
     AirPurifierComponent,
     LightBulbComponent,
     LightSensorComponent,